data_services/requirements.txt
paho-mqtt==1.6.1
requests==2.27.1
AWSIoTPythonSDK==1.5.2



